move duplicated stream reading functions from glvis.cpp and aux_js.cpp to stream_reader.{hc}pp
